Refactor manual CLI helpers
